requirements.txt
streamlit>=1.31.0
Pillow>=10.0.0
numpy>=1.24.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
tailwindcss>=3.4.0
plotly>=5.18.0
networkx>=3.2.0
openai>=1.12.0
transformers>=4.37.0
torch>=2.2.0
torchvision>=0.17.0
scikit-learn>=1.4.0
